requirements.txt
networkx
streamlit
pyyaml
matplotlib
